[refactor] split lineGeometry constructor function into a file
